Some refactoring, moving code
